Pulled slack rendering out of main path and only use it in slack case
